refact hexa

+++ ft_printf.c
@@ -8,13 +8,13 @@ static int print_fmt(char arg, va_list ap)
 	if (arg == 'c')
 		count += print_char(va_arg(ap, int));
 	if (arg == 'd')
-		count += print_decimal(va_arg(ap, signed long));
+		count += print_decimal(va_arg(ap, signed int));
 	if (arg == 's')
 		count += print_string(va_arg(ap, char *));
 	if (arg == 'p')
-		count += print_address(va_arg(ap, unsigned long));
+		count += print_address(va_arg(ap, unsigned int));
 	if (arg == 'i')
-		count += print_integer(va_arg(ap, signed long));
+		count += print_integer(va_arg(ap, signed int));
 	if (arg == 'u')
 		count += print_unsigned(va_arg(ap, unsigned int));
 	if (arg == 'x')
